+++ library/src/commonMain/kotlin/org/jraf/klibnotion/internal/model/block/Heading3BlockImpl.kt
@@ -27,11 +27,14 @@ package org.jraf.klibnotion.internal.model.block
 import org.jraf.klibnotion.model.base.UuidString
 import org.jraf.klibnotion.model.block.Block
 import org.jraf.klibnotion.model.block.Heading3Block
+import org.jraf.klibnotion.model.date.Timestamp
 import org.jraf.klibnotion.model.richtext.RichTextList
 
 internal data class Heading3BlockImpl(
     override val id: UuidString,
     override val text: RichTextList,
+    override val created: Timestamp,
+    override val lastEdited: Timestamp,
 ) : Heading3Block {
     override val children: List<Block>? = null
 }