+++ library/src/commonMain/kotlin/org/jraf/klibnotion/internal/model/block/ToDoBlockImpl.kt
@@ -27,10 +27,13 @@ package org.jraf.klibnotion.internal.model.block
 import org.jraf.klibnotion.model.base.UuidString
 import org.jraf.klibnotion.model.block.Block
 import org.jraf.klibnotion.model.block.ToDoBlock
+import org.jraf.klibnotion.model.date.Timestamp
 import org.jraf.klibnotion.model.richtext.RichTextList
 
 internal data class ToDoBlockImpl(
     override val id: UuidString,
+    override val created: Timestamp,
+    override val lastEdited: Timestamp,
     override val text: RichTextList,
     override val checked: Boolean,
     override var children: List<Block>?,